+++ app/src/main/java/com/techcourse/Application.java
@@ -1,10 +1,9 @@
 package com.techcourse;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.io.IOException;
 import java.util.stream.Stream;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class Application {
