+++ app/src/main/java/com/techcourse/controller/UserSession.java
@@ -1,9 +1,8 @@
 package com.techcourse.controller;
 
+import java.util.Optional;
 import com.techcourse.domain.User;
 import jakarta.servlet.http.HttpSession;
-
-import java.util.Optional;
 
 public class UserSession {
 
@@ -18,5 +17,6 @@ public class UserSession {
         return getUserFrom(session).isPresent();
     }
 
-    private UserSession() {}
+    private UserSession() {
+    }
 }
