+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/view/View.java
@@ -1,9 +1,8 @@
-package com.interface21.webmvc.servlet;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
+package com.interface21.webmvc.servlet.mvc.view;
 
 import java.util.Map;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 public interface View {
     void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
